Move some AtomType definitions to a new module.
