+++ opencog/haskell/opencog-atomspace.cabal
@@ -19,6 +19,7 @@ library
                    , OpenCog.AtomSpace.Types
                    , OpenCog.AtomSpace.Inheritance
                    , OpenCog.AtomSpace.Filter
+                   , OpenCog.AtomSpace.AtomType
 
   build-depends:     base         >=4.5
                    , transformers >=0.3
